98 fibo

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,24 +8,25 @@
 int main(void)
 {
 	int i = 0;
-	long j = 1, k = 2;
+	unsigned long long j = 1, k = 2;
 
 	while (i < 98)
 	{
 		if (i == 0)
-			printf("%ld", j);
+			printf("%llu", j);
 		else if (i == 1)
-			printf(", %ld", k);
+			printf(", %llu", k);
 		else
 		{
 			k += j;
 			j = k - j;
-			printf(", %ld", k);
+			printf(", %llu", k);
 		}
 
 		++i;
 	}
 
 	printf("\n");
+
 	return (0);
 }
